Move thingtalk templates to languages/thingtalk/
In preparation for supporting languages that are not thingtalk
