+++ languages/thingtalk/ast_manip.js
