+++ languages/thingtalk/common-parameters.genie
