+++ languages/thingtalk/en/aggregation.genie
