+++ languages/thingtalk/en/constants.genie
