+++ languages/thingtalk/en/filters.genie
