+++ languages/thingtalk/en/parameters.genie
