+++ languages/thingtalk/en/stream_tables.genie
