+++ languages/thingtalk/en/thingtalk.genie
