+++ languages/thingtalk/en/timers.genie
