+++ languages/thingtalk/utils.js
