+++ languages/thingtalk/zh-tw/constants.genie
