+++ languages/thingtalk/zh-tw/filters.genie
