+++ languages/thingtalk/zh-tw/parameters.genie
