+++ languages/thingtalk/zh-tw/thingtalk.genie
