+++ languages/thingtalk/zh-tw/timers.genie
