+++ lib/sentence-generator/index.js
@@ -26,7 +26,7 @@ const { clean, makeDummyEntities } = require('../utils');
 
 const $runtime = require('./runtime');
 const importGenie = require('../genie-compiler');
-const { typeToStringSafe } = require('../../languages/ast_manip');
+const { typeToStringSafe } = require('../../languages/thingtalk/ast_manip');
 const i18n = require('../i18n');
 
 function identity(x) {
