+++ test/test_sentence_generator.js
@@ -251,7 +251,7 @@ async function main() {
     const options = {
         rng: seedrandom.alea('almond is awesome'),
         locale: 'en-US',
-        templateFile: path.resolve(path.dirname(module.filename), '../languages/en/thingtalk.genie'),
+        templateFile: path.resolve(path.dirname(module.filename), '../languages/thingtalk/en/thingtalk.genie'),
         thingpediaClient: _tpClient,
         flags: {
             turking: false,
